Implemented --titledepth option.
This option allows the user to have the title be a partial fragment of
the fully qualified realpath to the source file being documented, with
the number of path components included specified as an integer argument
to --titledepth. This is highly useful when documenting projects that
have repeated/shared script names, thus requiring further path
components to uniquely identify a given script.

Three test cases have been added to validate the functionality of this
feature.