Implemented --prefix feature to allow shelldoc to set output file names programagically
This option allows the user to specify a prefix directory into which the
output file will be written. The output file name is the title (just the
part of the title wrapped in `` ``) with any instances of / replaced
with _. The intention is that this functionality can be used with Sphinx
to generate documents for a large number of scripts at once by simply
running shelldoc on each script via find -exec. This feature is intended
to be used in conjunction with --titledepth, but this is not required.

Due to the limits of the do_test.sh script as it stands, it is not
possible to provide adequate smoke testing for this feature at this
time.